pyproject.toml
[tool.poetry]
version = "0.1.0"
name = "build-chatbot"
description = "Build ChatBot"
authors = ["Nguyen Dinh Quoc Duy <nguyendqduy@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">= 3.11 < 3.12.0a1"

fastapi = "0.111.0"
uvicorn = { version = "0.30.6", extras = ["standard"] }
pydantic = "2.9.2"
python-multipart = "0.0.9"

Flask = "3.0.3"
Flask-Cors = "5.0.0"

passlib = { version = "1.7.4", extras = ["bcrypt"] }

requests = "2.32.3"
aiohttp = "3.10.8"
async-timeout = "*"

sqlalchemy = "2.0.32"
alembic = "1.13.2"
peewee = "3.17.6"
peewee-migrate = "1.12.2"
psycopg2-binary = "2.9.9"
PyMySQL = "1.1.1"
bcrypt = "4.2.0"

pymongo = "*"
redis = "*"
boto3 = "1.35.0"

argon2-cffi = "23.1.0"
APScheduler = "3.10.4"

openai = "0.28"
anthropic = "*"
google-generativeai = "0.7.2"
tiktoken = "*"

langchain = "0.2.15"
langchain-community = "0.2.12"
langchain-chroma = "0.1.4"

fake-useragent = "1.5.1"
chromadb = "0.5.9"
pymilvus = "2.4.7"

sentence-transformers = "3.2.0"
colbert-ai = "0.2.21"
einops = "0.8.0"

ftfy = "6.2.3"
pypdf = "4.3.1"
xhtml2pdf = "0.2.16"
pymdown-extensions = "10.11.2"
docx2txt = "0.8"
python-pptx = "1.0.0"
unstructured = "0.15.9"
nltk = "3.9.1"
Markdown = "3.7"
pypandoc = "1.13"
pandas = "2.2.3"
openpyxl = "3.1.5"
pyxlsb = "1.0.10"
xlrd = "2.0.1"
validators = "0.34.0"
psutil = "*"

opencv-python-headless = "4.10.0.84"
rapidocr-onnxruntime = "1.3.24"

fpdf2 = "2.7.9"
rank-bm25 = "0.2.2"

faster-whisper = "1.0.3"

PyJWT = { version = "2.9.0", extras = ["crypto"] }
authlib = "1.3.1"

black = "24.8.0"
langfuse = "2.44.0"
youtube-transcript-api = "0.6.2"
pytube = "15.0.0"

extract_msg = "*"
pydub = "*"
duckduckgo-search = "6.2.13"

docker = "7.1.0"
pytest = "8.3.2"
pytest-docker = "3.1.1"
pdfminer-six = "20221105"
pdf2image = "1.16.3"
pytesseract = "0.3.10"
vncorenlp = "1.0.3"
txtai = "7.5.1"
celery = "5.3.4"
weaviate-client = "4.9.3"
pre-commit = "4.0.1"
google-auth-oauthlib = "^1.2.1"
langgraph = "^0.2.53"
cohere = "^5.12.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
